Resolved the conflict between PlayerDashing.cs and PlayerMovement.cs
by merging them together
