+++ Assets/Levels/Prefabs/Player.prefab
@@ -13,7 +13,6 @@ GameObject:
   - component: {fileID: 8224430303388929554}
   - component: {fileID: 1698889524490382574}
   - component: {fileID: 3687398502885057350}
-  - component: {fileID: 1971792281472476401}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -162,21 +161,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 31a6c41f36f58974eb6f4c68b670e082, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  jumpHeight: 0
+  jumpHeight: 10
   size: 1
   maxSpeed: 10
   jumpForce: 10
---- !u!114 &1971792281472476401
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3404275623777843720}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 286963744f090964f99454567ddd7e78, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  dashDistance: 5
-  dashDuration: 0.5
+  dashingVelocity: 14
+  dashDistance: 2
+  dashingTime: 0.5
