--- Assets/Scripts/Player/PlayerDashing.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 286963744f090964f99454567ddd7e78
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
